+++ lib/src/integrations/test_coverage/command/dart/command/dart_test_coverage_command.dart
@@ -54,10 +54,17 @@ class DartTestCoverageCommand extends CliCommand<DartTestCoverageConfigData> {
       ]);
     }
 
+    final ignorePatterns = [
+      RegExp(r'.*\.g\.dart$'),
+      RegExp(r'.*\.navigator\.dart$'),
+      RegExp(r'.*/injectable\.config\.dart$'),
+    ];
+
     ImpaktfullCliLogger.startSpinner('Generating test coverage report...');
     final lcovFile = await testCoveragePlugin.testCoverage(
       path: '.',
       type: TestCoverageType.lcovInfo,
+      ignorePatterns: ignorePatterns,
     );
     ImpaktfullCliLogger.endSpinner();
     ImpaktfullCliLogger.log(lcovFile.printReport('Flutter'));