+++ lib/src/integrations/test_coverage/command/flutter/command/flutter_test_coverage_command.dart
@@ -38,10 +38,18 @@ class FlutterTestCoverageCommand
       ]);
       ImpaktfullCliLogger.endSpinner();
     }
+
+    final ignorePatterns = [
+      RegExp(r'.*\.g\.dart$'),
+      RegExp(r'.*\.navigator\.dart$'),
+      RegExp(r'.*/injectable\.config\.dart$'),
+    ];
+
     ImpaktfullCliLogger.startSpinner('Generating test coverage report...');
     final lcovFile = await testCoveragePlugin.testCoverage(
       path: '.',
       type: TestCoverageType.lcovInfo,
+      ignorePatterns: ignorePatterns,
     );
     ImpaktfullCliLogger.endSpinner();
     ImpaktfullCliLogger.log(lcovFile.printReport('Flutter'));